[refactor] Progress Bar component [optimize] Folder structure
