+++ package.json
@@ -49,7 +49,6 @@
         "@parcel/transformer-typescript-types": "~2.12.0",
         "@peculiar/webcrypto": "^1.5.0",
         "@tech_query/snabbdom-looks-like": "^2.0.1",
-        "@types/classnames": "^2.3.1",
         "@types/jest": "^29.5.12",
         "@types/resize-observer-browser": "^0.1.11",
         "@types/turndown": "^5.0.4",
