+++ source/Content/Card.tsx
@@ -1,8 +1,8 @@
 import classNames from 'classnames';
 import { FC, WebCellProps } from 'web-cell';
 
-import { Image, ImageProps } from './Image';
-import { TextColor, PositionY } from './type';
+import { Image, ImageProps } from '../Image';
+import { TextColor, PositionY } from '../type';
 
 export interface CardProps extends WebCellProps<HTMLDivElement> {
     bg?: TextColor;
