+++ source/Content/index.ts
@@ -1,9 +1,8 @@
-export * from './ListGroup';
 export * from './Table';
 export * from './Jumbotron';
 export * from './Card';
 export * from './MediaObject';
-export * from './EdgeDetector';
+export * from './ScrollBoundary';
 export * from './Collapse';
 export * from './Accordion';
-export * from './TabView';
+export * from './Tabs';
