+++ source/index.ts
@@ -1,23 +1,16 @@
 export * from './type';
 export * from './Ratio';
 export * from './Grid';
-export * from './Table';
-export * from './ScrollBoundary';
-export * from './Jumbotron';
-export * from './Card';
-export * from './MediaObject';
+export * from './Content';
 export * from './Form';
 export * from './Button';
 export * from './Reminder';
 export * from './Image';
 export * from './Prompt';
 export * from './Dropdown';
-export * from './Collapse';
-export * from './Accordion';
 export * from './Carousel';
 export * from './Nav';
 export * from './Navbar';
 export * from './Offcanvas';
-export * from './Tabs';
 export * from './CountDown';
 export * from './MonthCalendar';
